2nd Problem of AdventOfCode 2021
Couldn't think of anything clever on this one.  Just iterated through the list of readlines() and added and subtracted where needed based on where the number is.  Since down, up, and forward have different amount of characters, the indexes for the numbers are different.  I used len() for the "if" check but I think I may be able to just check the first character in the string.  Not sure which is faster or if it really has any significant difference.